+++ src/main/java/io/github/ititus/pdx/stellaris/user/save/RelationModifier.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 public class RelationModifier {
 
     public final boolean decay;
-    public final int value;
+    public final double value;
     public final String modifier;
     public final LocalDate startDate;
 
@@ -16,7 +16,7 @@ public class RelationModifier {
         PdxScriptObject o = s.expectObject();
         this.modifier = o.getString("modifier");
         this.startDate = o.getDate("start_date", null);
-        this.value = o.getInt("value");
+        this.value = o.getDouble("value");
         this.decay = o.getBoolean("decay", false);
     }
 }
